Reload pages in the browser when content has changed
Include a simple auto-reload script that frequently reads the hash
files. Reload the current page when content has changed.
Redirect to the home page when a has file is no longer available,
thus e.g. an error has occured while loading or parsing the DSL file
or when a software system has been renamed.

Inject the auto-reload script only when "serving" the
generated a web site but not when "generating" only.