Make element links in embedded SVGs relative
This fixes broken links when the generate site is not hosted as root
web site.
As a consequence we no longer generate embedded SVG's up-front,
but generate them as we need them with links relative to the link of
the current page.
To preserve generation speed for serve, we try to cache as much as we
can.