Hide internal properties
The Structurizr DSL library introduced element/relationship identifiers
as properties recently. Hide those for the site generation since we are
not really interested in presenting them to users.

Note that at the time of committing they still appear on the generated
diagrams, though that has already been addressed upstream in
com.structurizr:structurizr-export and will be fixed with a library
update once a new version has been released.