Write puml file after all other export formats
Since we use the existence of the puml file as up-to-date check,
we should write this after all other formats have been written.
This prevents a FileNotFoundException during concurent page
generation when the puml was already generated, thus we
assumed that the SVG has also been generated, but in fact the
SVG wasn't yet there.