Lower the JRE version requirement to version 17
Since we don't use any modern JRE features for this tool, there's no
need for it to always require the newest JRE to be installed. Therefore
it seems like a good idea to lower the requirement to the lowest JRE
version that is required by the libraries this tool uses. This happens
to be version 17.